QIpPacketHeader: correct the type of hopLimit member
In other places, it's referenced as int.

+++ src/network/socket/qabstractsocketengine_p.h
@@ -78,7 +78,7 @@ public:
     QHostAddress destinationAddress;
 
     uint ifindex;
-    qint16 hopLimit;
+    int hopLimit;
     quint16 senderPort;
     quint16 destinationPort;
 };
